[ERROR][CLEAN] - Refactor error handling and add new error system
[*] Move FSProblem enum in error::mod file

[*] rename file to correct typo
[+] Add ehanced error handling method based on error-stack librairy
