+++ src/error/standard.rs
@@ -1,4 +1,6 @@
 use crate::{Method, models::SecureDelete};
+use crate::error::FSProblem;
+
 #[cfg(not(feature = "error-stack"))]
 pub type Result<T> = core::result::Result<T, Error>;
 
@@ -11,32 +13,6 @@ pub enum Error {
     #[cfg(test)]
     FileCreationError(std::io::Error),
 }
-
-#[derive(Debug)]
-pub enum FSProblem {
-    Rename,
-    Opening,
-    Write,
-    Delete,
-    ReadFolder,
-    NotFound,
-    Permissions
-}
-
-impl core::fmt::Display for FSProblem {
-    fn fmt(&self, fmt: &mut core::fmt::Formatter) -> core::result::Result<(), core::fmt::Error> {
-        match self {
-            FSProblem::Rename => write!(fmt, "Rename"),
-            FSProblem::Opening => write!(fmt, "Opening"),
-            FSProblem::Write => write!(fmt, "writing"),
-            FSProblem::Delete => write!(fmt, "Delete"),
-            FSProblem::ReadFolder => write!(fmt, "Read Folder"),
-            FSProblem::NotFound => write!(fmt, "File/Folder not found"),
-            FSProblem::Permissions => write!(fmt, "Change permission error"),
-        }
-    }
-}
-
 
 impl core::fmt::Display for Error {
     fn fmt(&self, fmt: &mut core::fmt::Formatter) -> core::result::Result<(), core::fmt::Error> {
@@ -59,5 +35,4 @@ impl core::fmt::Display for Error {
     }
 }
 
-#[cfg(not(feature = "error-stack"))]
 impl std::error::Error for Error {}
