+++ src/methods/dod_522022_me.rs
@@ -30,7 +30,7 @@ mod std_test {
     use crate::tests::TestType;
     use crate::Method::Dod522022ME as EraseMethod;
     use crate::{Error, Result};
-    use crate::error::standard::FSProblem;
+    use crate::error::FSProblem;
 
     use pretty_assertions::{assert_eq, assert_ne};
