+++ src/methods/dod_522022_mece.rs
@@ -41,7 +41,7 @@ mod std_test {
     use crate::tests::TestType;
     use crate::Method::Dod522022MECE as EraseMethod;
     use crate::{Error, Result};
-    use crate::error::standard::FSProblem;
+    use crate::error::FSProblem;
 
     use pretty_assertions::{assert_eq, assert_ne};
