+++ src/methods/gutmann.rs
@@ -82,7 +82,7 @@ mod std_test {
     use crate::tests::TestType;
     use crate::Method::Gutmann as EraseMethod;
     use crate::{Error, Result};
-    use crate::error::standard::FSProblem;
+    use crate::error::FSProblem;
 
     use pretty_assertions::{assert_eq, assert_ne, assert_str_eq};
