+++ src/methods/hmgi_s5.rs
@@ -29,7 +29,7 @@ mod std_test {
     use crate::tests::TestType;
     use crate::Method::HmgiS5 as EraseMethod;
     use crate::{Error, Result};
-    use crate::error::standard::FSProblem;
+    use crate::error::FSProblem;
 
     use pretty_assertions::{assert_eq, assert_ne};
