+++ src/methods/pseudo_random.rs
@@ -22,7 +22,7 @@ mod std_test {
     use crate::tests::TestType;
     use crate::Method::PseudoRandom;
     use crate::{Error, Result};
-    use crate::error::standard::FSProblem;
+    use crate::error::FSProblem;
 
     use pretty_assertions::{assert_eq, assert_ne};
