+++ src/methods/rcmp_tssit_ops_ii.rs
@@ -33,8 +33,7 @@ mod std_test {
     use crate::tests::TestType;
     use crate::Method::RcmpTssitOpsII as EraseMethod;
     use crate::{Error, Result};
-    use crate::error::standard::FSProblem;
-
+    use crate::error::FSProblem;
     use pretty_assertions::{assert_eq, assert_ne};
 
     const METHOD_NAME: &str = "rcmp_tssit_ops_ii";
